Introduce MetadatastoreUtils
