+++ languages/PSBgen.groovy
@@ -7,6 +7,7 @@ import groovy.transform.*
 // define script properties
 @Field BuildProperties props = BuildProperties.getInstance()
 @Field def buildUtils= loadScript(new File("${props.zAppBuildDir}/utilities/BuildUtilities.groovy"))
+@Field def metadataUtils= loadScript(new File("MetadatastoreUtilities.groovy"))
 
 println("** Building ${argMap.buildList.size()} ${argMap.buildList.size() == 1 ? 'file' : 'files'} mapped to ${this.class.getName()}.groovy script")
 
@@ -52,7 +53,7 @@ sortedList.each { buildFile ->
 		String errorMsg = "*! The assembly return code ($rc) for $buildFile exceeded the maximum return code allowed ($maxRC)"
 		println(errorMsg)
 		props.error = "true"
-		buildUtils.updateBuildResult(errorMsg:errorMsg,logs:["${member}.log":logFile])
+		metadataUtils.updateBuildResult(errorMsg:errorMsg,logs:["${member}.log":logFile])
 	}
 	else {
 
@@ -63,7 +64,7 @@ sortedList.each { buildFile ->
 			String errorMsg = "*! The link edit return code ($rc) for $buildFile exceeded the maximum return code allowed ($maxRC)"
 			println(errorMsg)
 			props.error = "true"
-			buildUtils.updateBuildResult(errorMsg:errorMsg,logs:["${member}.log":logFile])
+			metadataUtils.updateBuildResult(errorMsg:errorMsg,logs:["${member}.log":logFile])
 		}
 
 		else{
@@ -75,7 +76,7 @@ sortedList.each { buildFile ->
 					String errorMsg = "*! The acbgen return code ($rc) for $buildFile exceeded the maximum return code allowed ($maxRC)"
 					println(errorMsg)
 					props.error = "true"
-					buildUtils.updateBuildResult(errorMsg:errorMsg,logs:["${member}.log":logFile])
+					metadataUtils.updateBuildResult(errorMsg:errorMsg,logs:["${member}.log":logFile])
 				}
 			}
 		}
