+++ utilities/ImpactUtilities.groovy
@@ -14,6 +14,7 @@ import com.ibm.dbb.dependency.internal.*
 @Field BuildProperties props = BuildProperties.getInstance()
 @Field def gitUtils= loadScript(new File("GitUtilities.groovy"))
 @Field def buildUtils= loadScript(new File("BuildUtilities.groovy"))
+@Field def metadataUtils= loadScript(new File("MetadatastoreUtilities.groovy"))
 @Field def dependencyScannerUtils= loadScript(new File("DependencyScannerUtilities.groovy"))
 @Field String hashPrefix = ':githash:'
 
@@ -48,7 +49,7 @@ def createImpactBuildList() {
 	}
 
 	// scan files and update source collection for impact analysis
-	updateCollection(changedFiles, deletedFiles, renamedFiles)
+	metadataUtils.updateCollection(changedFiles, deletedFiles, renamedFiles)
 
 
 	if (calculatedChanges) {
@@ -112,7 +113,7 @@ def createImpactBuildList() {
 						}
 					} else {
 						String warningMsg = "*! $impactFile is impacted by changed file $changedFile, but is not added to build list, because it is not mapped to a language script."
-						buildUtils.updateBuildResult(warningMsg:warningMsg)
+						metadataUtils.updateBuildResult(warningMsg:warningMsg)
 						println(warningMsg)
 					}
 				}
@@ -188,7 +189,7 @@ def createMergeBuildList(){
 	(changedFiles, deletedFiles, renamedFiles, changedBuildProperties) = calculateChangedFiles(null)
 
 	// scan files and update source collection
-	updateCollection(changedFiles, deletedFiles, renamedFiles)
+	metadataUtils.updateCollection(changedFiles, deletedFiles, renamedFiles)
 
 	// iterate over changed file and add them to the buildSet
 
@@ -485,7 +486,7 @@ def scanOnlyStaticDependencies(List buildList){
 				if ((isLinkEdited && isLinkEdited.toBoolean()) || scriptMapping == "LinkEdit.groovy"){
 					try{
 						if (props.verbose) println ("*** Scanning load module $loadPDSMember of $buildFile")
-						saveStaticLinkDependencies(buildFile, props."${langPrefix}_loadPDS", logicalFile)
+						metadataUtils.saveStaticLinkDependencies(buildFile, props."${langPrefix}_loadPDS", logicalFile)
 					}
 					catch (com.ibm.dbb.build.ValidationException e){
 						println ("!* Error scanning output file for $buildFile  : $loadPDSMember")
@@ -502,211 +503,6 @@ def scanOnlyStaticDependencies(List buildList){
 	}
 }
 
-
-
-
-
-def updateCollection(changedFiles, deletedFiles, renamedFiles) {
-
-	if (!MetadataStoreFactory.metadataStoreExists()) {
-		if (props.verbose) println "** Unable to update collections. No Metadata Store."
-		return
-	}
-	MetadataStore metadataStore = MetadataStoreFactory.getMetadataStore()
-
-	if (props.verbose) println "** Updating collections ${props.applicationCollectionName} and ${props.applicationOutputsCollectionName}"
-	//def scanner = new DependencyScanner()
-	List<LogicalFile> logicalFiles = new ArrayList<LogicalFile>()
-	List<PathMatcher> excludeMatchers = buildUtils.createPathMatcherPattern(props.excludeFileList)
-
-	verifyCollections()
-
-	// remove deleted files from collection
-	deletedFiles.each { file ->
-		// files in a collection are stored as relative paths from a source directory
-		if (props.verbose) println "*** Deleting logical file for $file"
-		logicalFile = buildUtils.relativizePath(file)
-		metadataStore.getCollection(props.applicationCollectionName).deleteLogicalFile(logicalFile)
-		metadataStore.getCollection(props.applicationOutputsCollectionName).deleteLogicalFile(logicalFile)
-	}
-
-	// remove renamed files from collection
-	renamedFiles.each { file ->
-		// files in a collection are stored as relative paths from a source directory
-		if (props.verbose) println "*** Deleting renamed logical file for $file"
-		logicalFile = buildUtils.relativizePath(file)
-		metadataStore.getCollection(props.applicationCollectionName).deleteLogicalFile(logicalFile)
-		metadataStore.getCollection(props.applicationOutputsCollectionName).deleteLogicalFile(logicalFile)
-	}
-
-	if (props.createTestcaseDependency && props.createTestcaseDependency.toBoolean() && changedFiles && changedFiles.size() > 1) {
-		changedFiles = sortFileList(changedFiles);
-		if (props.verbose) println "*** Sorted list of changed files: $changedFiles"
-	}
-
-	// scan changed files
-	changedFiles.each { file ->
-
-		// make sure file is not an excluded file
-		if ( new File("${props.workspace}/${file}").exists() && !buildUtils.matches(file, excludeMatchers)) {
-			// files in a collection are stored as relative paths from a source directory
-
-			def scanner = dependencyScannerUtils.getScanner(file)
-			try {
-				def logicalFile
-				if (scanner != null) {
-					if (props.verbose) println "*** Scanning file $file (${props.workspace}/${file} with ${scanner.getClass()})"
-					logicalFile = scanner.scan(file, props.workspace)
-				} else {
-					// The below logic should be replaced with Registration Scanner when available
-					// See reported idea: https://ibm-z-software-portal.ideas.ibm.com/ideas/DBB-I-48
-					if (props.verbose) println "*** Skipped scanning file $file (${props.workspace}/${file})"
-					
-					// New logical file with Membername, buildfile, language set to file extension
-					logicalFile = new LogicalFile(CopyToPDS.createMemberName(file), file, file.substring(file.lastIndexOf(".") + 1).toUpperCase(), false, false, false)
-					
-					// Add logicalFile to LogicalFileCache
-					LogicalFileCache.add(props.workspace, logicalFile)
-				}
-				if (props.verbose) println "*** Logical file for $file =\n$logicalFile"
-
-				// Update logical file with dependencies to build properties
-				if (props.impactBuildOnBuildPropertyChanges && props.impactBuildOnBuildPropertyChanges.toBoolean()){
-					createPropertyDependency(file, logicalFile)
-				}
-
-				// If configured, update test case program dependencies
-				if (props.createTestcaseDependency && props.createTestcaseDependency.toBoolean()) {
-					// If the file is a zUnit configuration file (BZUCFG)
-					if (scanner != null && scanner.getClass() == com.ibm.dbb.dependency.ZUnitConfigScanner) {
-
-						def logicalDependencies = logicalFile.getLogicalDependencies()
-
-						def sysTestDependency = logicalDependencies.find{it.getLibrary().equals("SYSTEST")} // Get the test case program from testcfg
-						def sysProgDependency = logicalDependencies.find{it.getLibrary().equals("SYSPROG")} // Get the application program name from testcfg
-
-						if (sysTestDependency){
-							// find in local list of logical files first (batch processing)
-							def testCaseFiles = logicalFiles.findAll{it.getLname().equals(sysTestDependency.getLname())}
-							if (!testCaseFiles){ // alternate retrieve it from the collection
-								testCaseFiles = metadataStore.getCollection(props.applicationCollectionName).getLogicalFiles(sysTestDependency.getLname()).find{
-									it.getLanguage().equals("COB")
-								}
-							}
-							testCaseFiles.each{
-								it.addLogicalDependency(new LogicalDependency(sysProgDependency.getLname(),"SYSPROG","PROGRAMDEPENDENCY"))
-								if (props.verbose) println "*** Updating dependencies for test case program ${it.getFile()} =\n$it"
-								logicalFiles.add(it)
-							}
-						}
-					}
-				}
-
-				logicalFiles.add(logicalFile)
-
-			} catch (Exception e) {
-
-				String warningMsg = "***** Scanning failed for file $file (${props.workspace}/${file})"
-				buildUtils.updateBuildResult(warningMsg:warningMsg)
-				println(warningMsg)
-				e.printStackTrace()
-
-				// terminate when continueOnScanFailure is not set to true
-				if(!(props.continueOnScanFailure == 'true')){
-					println "***** continueOnScan Failure set to false. Build terminates."
-					System.exit(1)
-				}
-			}
-
-			// save logical files in batches of 500 to avoid running out of heap space
-			if (logicalFiles.size() == 500) {
-				if (props.verbose)
-					println "** Storing ${logicalFiles.size()} logical files in MetadataStore collection '$props.applicationCollectionName'"
-				metadataStore.getCollection(props.applicationCollectionName).addLogicalFiles(logicalFiles)
-				logicalFiles.clear()
-			}
-		}
-	}
-
-	// save logical files
-	if (props.verbose)
-		println "** Storing ${logicalFiles.size()} logical files in MetadataStore collection '$props.applicationCollectionName'"
-	metadataStore.getCollection(props.applicationCollectionName).addLogicalFiles(logicalFiles)
-	
-}
-
-/*
- * saveStaticLinkDependencies - Scan the load module to determine LINK dependencies. Impact resolver can use
- * these to determine that this file gets rebuilt if a LINK dependency changes.
- */
-def saveStaticLinkDependencies(String buildFile, String loadPDS, LogicalFile logicalFile) {
-	MetadataStore metadataStore = MetadataStoreFactory.getMetadataStore()
-	if (metadataStore && !props.error && !props.preview) {
-		LinkEditScanner scanner = new LinkEditScanner()
-		if (props.verbose) println "*** Scanning load module for $buildFile"
-		LogicalFile scannerLogicalFile = scanner.scan(buildUtils.relativizePath(buildFile), loadPDS)
-		if (props.verbose) println "*** Logical file = \n$scannerLogicalFile"
-
-		// overwrite original logicalDependencies with load module dependencies
-		logicalFile.setLogicalDependencies(scannerLogicalFile.getLogicalDependencies())
-
-		// Store logical file and indirect dependencies to the outputs collection
-		metadataStore.getCollection("${props.applicationOutputsCollectionName}").addLogicalFile( logicalFile );
-	}
-}
-
-/*
- * verifyCollections - verifies that the application collections exists. If not it will
- * create or clone the collections.
- * Uses build properties
- */
-def verifyCollections() {
-	MetadataStore metadataStore = MetadataStoreFactory.getMetadataStore()
-	if (!metadataStore) {
-		if (props.verbose) println "** Unable to verify collections. No metadata store."
-		return
-	}
-
-	String mainCollectionName = "${props.application}-${props.mainBuildBranch}"
-	String mainOutputsCollectionName = "${props.application}-${props.mainBuildBranch}-outputs"
-
-	// check source collection
-	if (!metadataStore.collectionExists(props.applicationCollectionName)) {
-		if (props.topicBranchBuild) {
-			if (metadataStore.collectionExists(mainCollectionName)) {
-				metadataStore.copyCollection(mainCollectionName, props.applicationCollectionName)
-				if (props.verbose) println "** Cloned collection ${props.applicationCollectionName} from $mainCollectionName"
-			}
-			else {
-				metadataStore.createCollection(props.applicationCollectionName)
-				if (props.verbose) println "** Created collection ${props.applicationCollectionName}"
-			}
-		}
-		else {
-			metadataStore.createCollection(props.applicationCollectionName)
-			if (props.verbose) println "** Created collection ${props.applicationCollectionName}"
-		}
-	}
-
-	// check outputs collection
-	if (!metadataStore.collectionExists(props.applicationOutputsCollectionName)) {
-		if (props.topicBranchBuild) {
-			if (metadataStore.collectionExists(mainOutputsCollectionName)) {
-				metadataStore.copyCollection("${mainOutputsCollectionName}", props.applicationOutputsCollectionName)
-				if (props.verbose) println "** Cloned collection ${props.applicationOutputsCollectionName} from $mainOutputsCollectionName"
-			}
-			else {
-				metadataStore.createCollection(props.applicationOutputsCollectionName)
-				if (props.verbose) println "** Created collection ${props.applicationOutputsCollectionName}"
-			}
-		}
-		else {
-			metadataStore.createCollection(props.applicationOutputsCollectionName)
-			if (props.verbose) println "** Created collection ${props.applicationOutputsCollectionName}"
-		}
-	}
-
-}
 
 /*
  *  calculates the correct filepath from the git diff, due to different offsets in the directory path
